all files

+++ HashTable.cpp
@@ -3,30 +3,35 @@
 #include <random>
 #include <stdexcept>
 
-// Initialize the static constant
+// Initialize the static constant - ADDED BACK
 const size_t HashTable::DEFAULT_INITIAL_CAPACITY;
 
 // ============================================================================
 // HashTableBucket CLASS IMPLEMENTATION
 // ============================================================================
 
+// Default constructor - sets bucket type to ESS
 HashTableBucket::HashTableBucket() : key(""), value(0), type(BucketType::ESS) {}
 
-HashTableBucket::HashTableBucket(string key, size_t value)
+// Parameterized constructor - initializes key, value and sets type to NORMAL
+HashTableBucket::HashTableBucket(string key, int value)
     : key(key), value(value), type(BucketType::NORMAL) {}
 
-void HashTableBucket::load(string newKey, size_t newValue) {
+// Load method - loads key-value pair and marks bucket as NORMAL
+void HashTableBucket::load(string newKey, int newValue) {
     key = newKey;
     value = newValue;
     type = BucketType::NORMAL;
 }
 
+// Clear the bucket - mark as Empty After Remove
 void HashTableBucket::clear() {
     key = "";
     value = 0;
     type = BucketType::EAR;
 }
 
+// Check if bucket is empty (regardless of ESS/EAR state)
 bool HashTableBucket::isEmpty() const {
     return type == BucketType::ESS || type == BucketType::EAR;
 }
@@ -43,11 +48,12 @@ bool HashTableBucket::isNormal() const {
     return type == BucketType::NORMAL;
 }
 
+// Getters
 string HashTableBucket::getKey() const {
     return key;
 }
 
-size_t HashTableBucket::getValue() const {
+int HashTableBucket::getValue() const {
     return value;
 }
 
@@ -55,10 +61,12 @@ BucketType HashTableBucket::getType() const {
     return type;
 }
 
-void HashTableBucket::setValue(size_t newValue) {
+// Setter
+void HashTableBucket::setValue(int newValue) {
     value = newValue;
 }
 
+// Stream insertion operator for HashTableBucket
 ostream& operator<<(ostream& os, const HashTableBucket& bucket) {
     if (bucket.isNormal()) {
         os << "<" << bucket.key << ", " << bucket.value << ">";
@@ -70,11 +78,13 @@ ostream& operator<<(ostream& os, const HashTableBucket& bucket) {
 // HashTable CLASS IMPLEMENTATION
 // ============================================================================
 
+// Constructor - takes initial capacity, defaults to 8
 HashTable::HashTable(size_t initCapacity) : numItems(0) {
     tableData.resize(initCapacity);
     generateOffsets(initCapacity);
 }
 
+// Hash function
 size_t HashTable::hashFunction(const string& key) const {
     size_t hash = 0;
     for (char c : key) {
@@ -83,6 +93,7 @@ size_t HashTable::hashFunction(const string& key) const {
     return hash % tableData.size();
 }
 
+// Generate pseudo-random probing offsets
 void HashTable::generateOffsets(size_t size) {
     offsets.clear();
     for (size_t i = 1; i < size; i++) {
@@ -93,6 +104,7 @@ void HashTable::generateOffsets(size_t size) {
     shuffle(offsets.begin(), offsets.end(), g);
 }
 
+// Resize table if load factor reaches threshold
 void HashTable::resizeIfNeeded() {
     if (alpha() >= 0.5) {
         vector<HashTableBucket> oldTable = tableData;
@@ -124,6 +136,7 @@ size_t HashTable::findKeyIndex(const string& key) const {
     for (size_t i = 0; i < offsets.size(); i++) {
         currentIndex = (home + offsets[i]) % tableData.size();
 
+        // Stop if we find a never-used bucket
         if (tableData[currentIndex].isEmptySinceStart()) {
             return tableData.size(); // Not found
         }
@@ -136,12 +149,13 @@ size_t HashTable::findKeyIndex(const string& key) const {
     return tableData.size(); // Not found
 }
 
-bool HashTable::insert(string key, size_t value) {
+// Insert key-value pair - returns true if successful, false for duplicates
+bool HashTable::insert(string key, int value) {
     resizeIfNeeded();
 
-    // Check for duplicate key
+    // Check for duplicate key (no duplicates allowed)
     if (contains(key)) {
-        return false;
+        return false; // Insertion failed - duplicate key
     }
 
     size_t home = hashFunction(key);
@@ -151,7 +165,7 @@ bool HashTable::insert(string key, size_t value) {
     if (tableData[home].isEmpty()) {
         tableData[home].load(key, value);
         numItems++;
-        return true;
+        return true; // Successfully inserted
     }
 
     // Use probing sequence to find empty slot
@@ -161,40 +175,45 @@ bool HashTable::insert(string key, size_t value) {
         if (tableData[currentIndex].isEmpty()) {
             tableData[currentIndex].load(key, value);
             numItems++;
-            return true;
+            return true; // Successfully inserted
         }
     }
 
-    return false;
+    return false; // Table is full (shouldn't happen with proper resizing)
 }
 
+// Remove key-value pair - marks bucket as empty-after-remove
 bool HashTable::remove(string key) {
     size_t index = findKeyIndex(key);
     if (index < tableData.size()) {
-        tableData[index].clear();
+        tableData[index].clear(); // Mark as EAR
         numItems--;
-        return true;
+        return true; // Successfully removed
     }
-    return false;
+    return false; // Key not found
 }
 
+// Check if key exists in table
 bool HashTable::contains(const string& key) const {
     return findKeyIndex(key) < tableData.size();
 }
 
-optional<size_t> HashTable::get(const string& key) const {
+// Get value associated with key - returns optional<int>
+optional<int> HashTable::get(const string& key) const {
     size_t index = findKeyIndex(key);
     if (index < tableData.size()) {
-        return tableData[index].getValue();
+        return tableData[index].getValue(); // Returns int
     }
-    return nullopt;
+    return nullopt; // Key not found
 }
 
-// Fixed operator[] implementation
-size_t& HashTable::operator[](const string& key) {
-    // First, ensure the key exists (insert if it doesn't)
+// Bracket operator - returns reference to value, allows assignment
+int& HashTable::operator[](const string& key) {
+    // For our implementation, we'll insert if key doesn't exist
+    // (The requirement says undefined behavior is acceptable for missing keys,
+    // but we choose to be more robust)
     if (!contains(key)) {
-        insert(key, 0);
+        insert(key, 0); // Insert with default value 0
     }
 
     // Find the key's index
@@ -206,11 +225,12 @@ size_t& HashTable::operator[](const string& key) {
     }
 
     // Fallback - should never reach here if insert worked
-    static size_t dummy;
+    static int dummy;
     dummy = 0;
     return dummy;
 }
 
+// Returns vector with all keys currently in the table
 vector<string> HashTable::keys() const {
     vector<string> keyList;
     for (size_t i = 0; i < tableData.size(); i++) {
@@ -221,19 +241,23 @@ vector<string> HashTable::keys() const {
     return keyList;
 }
 
+// Returns current load factor (size/capacity)
 double HashTable::alpha() const {
     if (tableData.size() == 0) return 0.0;
     return static_cast<double>(numItems) / static_cast<double>(tableData.size());
 }
 
+// Returns total number of buckets in the table
 size_t HashTable::capacity() const {
     return tableData.size();
 }
 
+// Returns number of key-value pairs in the table
 size_t HashTable::size() const {
     return numItems;
 }
 
+// Stream insertion operator for HashTable - only prints occupied buckets
 ostream& operator<<(ostream& os, const HashTable& hashTable) {
     bool foundItems = false;
     for (size_t i = 0; i < hashTable.tableData.size(); i++) {
